1) Applied Peter's patch to support 16 bit windows (and remove the "ssgInit called    without a valid OpenGL context" error, instead a better error message is printed. 2) Changed sdldrv to follow our layout, added some documentation, fixed a few    warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@2373 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/race_manager.cpp
@@ -355,7 +355,7 @@ void RaceManager::RaceFinished(const Kart *kart, float time)
     }
 
     assert(pos-1 >= 0);
-    assert(pos-1 < m_kart_status.size());
+    assert(pos-1 < (int)m_kart_status.size());
 
     m_kart_status[i].m_score        += m_score_for_position[pos-1];
     m_kart_status[i].m_last_score    = m_score_for_position[pos-1];
